moved Matching game to router
