+++ client/app/(tabs)/index.tsx
@@ -4,16 +4,14 @@ import { Colors } from '@/constants/Colors';
 import { WelcomeHeader } from '@/components/WelcomeHeader';
 import { BrainExperienceBar } from '@/components/BrainExperienceBar';
 import { GameButtonLayout } from '@/components/GameButtonLayout';
-import { Matching } from '@/components/views/Matching';
 
 export default function HomeScreen() {
   return (
-    <Matching />
-    // <ThemedView style={styles.page} lightColor={Colors.primary} darkColor={Colors.primary}>
-    //   <WelcomeHeader />
-    //   <BrainExperienceBar />
-    //   <GameButtonLayout />
-    // </ThemedView>
+    <ThemedView style={styles.page} lightColor={Colors.primary} darkColor={Colors.primary}>
+      <WelcomeHeader />
+      <BrainExperienceBar />
+      <GameButtonLayout />
+    </ThemedView>
   );
 }
